pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "corvus-star"
version = "0.1.0"
description = "Sovereign AI Framework"
readme = "README.md"
authors = [
    { name = "Craig", email = "craig@example.com" }
]
requires-python = ">=3.10"
dependencies = [
    "requests",
    "playwright",
    "pytest",
    "ruff",
    "vulture",
    "psutil",
    "python-dotenv",
    "PyYAML",
    "textual",
    "pytest-asyncio",
    "google-genai",
    "google-generativeai",
    "colorama",
    "radon",
    "chromadb"
]

[project.scripts]
cstar = "src.core.cstar_dispatcher:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
    "mock_project",
    "sterileAgent",
    ".corvus_quarantine",
    ".agent/chroma_db",
    ".agent/scripts",
    ".agent/prompts",
    ".agent/workflows",
    "tests/empire_tests/src/error_real.py",
    "tests/gauntlet",
    "tests/fixtures",
    "tmp_diag",
    "bin"
]

[tool.ruff.lint]
select = [
    "E", "W", "F", "B", "C4", "UP", "SIM", "I", "N", "ANN", "S", "PTH", "RUF"
]
ignore = [
    "ANN204"
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101", "ANN"]
